+++ lib/hypothesis/errors.rb
@@ -1,0 +1,10 @@
+module Hypothesis
+  class Error < Exception
+  end
+
+  class UnsatisfiedAssumption < Error
+  end
+
+  class UsageError < Error
+  end
+end
